% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_grid_size.R
\name{as_grid_size}
\alias{as_grid_size}
\title{Turn a \code{loon} size to a \code{grid} size}
\usage{
as_grid_size(
  size,
  type = c("points", "texts", "images", "radial", "parallel", "polygon", "lines"),
  adjust = 1,
  ...
)
}
\arguments{
\item{size}{input \code{loon} size}

\item{type}{glyph type; one of "points", "texts", "images",
"radial", "parallel", "polygon", "lines".}

\item{adjust}{a pixel (px) at 96\code{DPI} (dots per inch) is equal to 0.75 point.
However, for different machines, the \code{DPI} is slightly different.
Argument \code{adjust} is used to twist the size. IT IS A HACK and should be removed
in the later version.}

\item{...}{some arguments used to specify the size, e.g. \code{pch} for "points",
\code{ratio} for "image" and \code{p} for "parallel".}
}
\description{
The size of \code{loon} is determined by pixel (px), while, in
\code{grid} graphics, the size is determined by pointsize (pt)
}
