% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ColorList.R
\name{as_hex6color}
\alias{as_hex6color}
\title{Return a 6 hexidecimal digit color representations}
\usage{
as_hex6color(color)
}
\arguments{
\item{color}{input color}
}
\description{
Return a 6 hexidecimal digit color representations
}
\details{
Compared with \code{hex12tohex6()}, it could accommodate 6 digit code, 12 digit code or
real color names.
}
\examples{
color <- c("#FF00FF", "#999999999999", "red")
# return 12 hexidecimal digit color
loon:::l_hexcolor(color)
# return 6 hexidecimal digit color
as_hex6color(color)
# return color names
l_colorName(color)

\dontrun{# WRONG COLORS
hex12tohex6(color)}
}
\seealso{
\code{\link{l_hexcolor}}, \code{\link{hex12tohex6}},
\code{\link{l_colorName}}
}
