% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_serialaxes.R
\name{char2num.data.frame}
\alias{char2num.data.frame}
\title{A Character Data Frame to a Numerical Data Frame}
\usage{
char2num.data.frame(chardataframe)
}
\arguments{
\item{chardataframe}{A char data frame}
}
\description{
Turn a data frame of characters to a data frame
of numerical values. If the character cannot be converted to
numerical in direct, it will be turned to factor first, then to numerical data
}
\examples{
data <- data.frame(x = c("1", "2", "3"),
                   y = c("foo", "bar", "foo"),
                   z = 4:6)
# ERROR
# data + 1
numData <- char2num.data.frame(data)
numData + 1

if(interactive()) {
  s <- l_serialaxes(iris)
  data <- s["data"]
  # it is a character data frame
  data[1,1]
  numData <- char2num.data.frame(data)
  numData[1,1]
}
}
