% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphutils.R
\name{complement.loongraph}
\alias{complement.loongraph}
\title{Create the Complement Graph of a loon Graph}
\usage{
\method{complement}{loongraph}(x)
}
\arguments{
\item{x}{loongraph object}
}
\value{
graph object of class loongraph
}
\description{
Creates a complement graph of a graph
}
\details{
This method is currently only implemented for undirected graphs.
}
