% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphutils.R
\name{complement}
\alias{complement}
\title{Create the Complement Graph of a Graph}
\usage{
complement(x)
}
\arguments{
\item{x}{graph or loongraph object}
}
\value{
graph object
}
\description{
Creates a complement graph of a graph
}
