% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBinInfo.R, R/l_binCut.R
\name{Extract-Bin-Info}
\alias{Extract-Bin-Info}
\alias{getBinData}
\alias{l_getBinIds}
\alias{l_breaks}
\alias{l_binCut}
\title{Some useful \code{loon} histogram functions}
\usage{
getBinData(widget)

l_getBinIds(widget)

l_breaks(widget)

l_binCut(widget, labels = NULL, digits = 2)
}
\arguments{
\item{widget}{A \code{loon} \code{l_hist} widget.}

\item{labels}{Labels for the levels of the resulting category.
By default, labels are constructed using "(a,b]" interval notation.
If labels = FALSE, simple integer codes are returned instead of a factor.
If labels are recognized as colors, character will be returned directly.}

\item{digits}{The number of digits used in formatting the breaks.}
}
\description{
Here list several useful functions for \code{loon} histograms.
}
\details{
\code{getBinData} is utilized to extract \code{loon} bin information,
such as the counts, locations, points id of each colored bin.
All rest useful functions in this page are built based on it.

\code{l_getBinIds} returns the ids of each bin

\code{l_breaks} returns the range (breaks) of each bin

\code{l_binCut} divides \code{l_hist} widget \code{x} into current histogram intervals and codes values
\code{x} according to which interval they fall. It is very similar to \code{\link{cut}} in \code{base} package.
}
\examples{
# getBinData
p <- l_hist(mtcars$mpg)
getBinData(p)


# l_binCut
p <- l_hist(iris)
binCut <- l_binCut(p)
p['color'] <- binCut
## ggplot color hue
gg_color_hue <- function(n) {
  hues <- seq(15, 375, length = n + 1)
  hcl(h = hues, l = 65, c = 100)[1:n]
}

# l_getBinIds
## number of bins
binIds <- l_getBinIds(p)
p['color'] <- l_binCut(p, labels = gg_color_hue(length(binIds)))

## The count of each bin
lengths(binIds)
# l_breaks
breaks <- l_breaks(p)
}
