% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_layout.R
\name{facet_separate_layout}
\alias{facet_separate_layout}
\title{layout separately}
\usage{
facet_separate_layout(
  plots,
  subtitles,
  title = "",
  xlabel = "",
  ylabel = "",
  sep = "*",
  maxCharInOneRow = 10,
  ...
)
}
\arguments{
\item{plots}{A list of \code{loon} plots}

\item{subtitles}{The subtitles of the layout. It is a list and the length is equal to
the number of \code{by} variables. Each element in a list is the unique values of such \code{by} variable.}

\item{title}{The title of the widget}

\item{xlabel}{The xlabel of the widget}

\item{ylabel}{The ylabel of the widget}

\item{sep}{The character string to separate or combine a vector}

\item{maxCharInOneRow}{deprecated}

\item{...}{named arguments to modify plot states.
See \code{\link{l_info_states}} of any instantiated l_plot for examples of names and values.}
}
\description{
layout separately
}
