% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facets.R
\name{get_facets}
\alias{get_facets}
\title{Creates a loon plot for each facet from an existing loon plot.}
\usage{
get_facets(widget, ...)
}
\arguments{
\item{widget}{the loon widget}

\item{...}{other arguments to the function used to create
the loon plot for each facet.  Depending on the plot being facetted, these
include the arguments \code{by},  \code{on}, \code{parent},  \code{layout},
\code{inheritLayers}, \code{separate}, and \code{bySubstitute}.
See \code{\link{l_facet}} for details on these parameters.
Other loon plot constructor parameters can be passed as well this way.}
}
\value{
A list containing the named components \code{plots}, \code{subtitles},
\code{child = child}, and \code{new.toplevel} containing the facets as plots
and other relevant information to construct the facetted plot.
}
\description{
A generic function used by \code{l_facet} when facetting an existing loon plot.
}
\seealso{
\code{\link{l_facet}}
}
\keyword{internal}
