% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob.R
\name{get_font_info_from_tk}
\alias{get_font_info_from_tk}
\title{Return Font Information}
\usage{
get_font_info_from_tk(tkFont)
}
\arguments{
\item{tkFont}{A specified tk font character, one of
\code{l_getOption("font-scales")}, \code{l_getOption("font-title")},
\code{l_getOption("font-xlabel")}, \code{l_getOption("font-ylabel")}}
}
\value{
A list of font information, containing font "family",
font "face" and font "size"
}
\description{
Return Font Information
}
\examples{
fontscales <- l_getOption("font-scales")
get_font_info_from_tk(fontscales)
}
