% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob.R
\name{glyph_to_pch}
\alias{glyph_to_pch}
\title{Glyph to Pch}
\usage{
glyph_to_pch(glyph)
}
\arguments{
\item{glyph}{glyph type in \code{loon}, must be "circle", "ocircle",
"ccircle", "square", "osquare", "csquare", "triangle", "otriangle", "ctriangle",
"diamond", "cdiamond", "odiamond". If the input glyph is not valid, \code{NA} is returned.}
}
\value{
a \code{\link{pch}} type
}
\description{
turn a loon point glyph to an R \code{graphics} plotting 'character' (pch)
}
\examples{
glyph_to_pch(c("circle", "ocircle", "ccircle",
               "square", "osquare", "csquare",
               "triangle", "otriangle", "ctriangle",
               "diamond", "cdiamond", "odiamond",
               "foo"))
}
