% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphutils.R
\name{graphreduce}
\alias{graphreduce}
\title{Make each space in a node apprear only once}
\usage{
graphreduce(graph, separator)
}
\arguments{
\item{graph}{graph of class loongraph}

\item{separator}{one character that separates the spaces in node names}
}
\value{
graph object of class loongraph
}
\description{
Reduce a graph to have unique node names
}
\details{
Note this is a string based operation. Node names must not contain
  the separator character!
}
\examples{
G <- completegraph(nodes=LETTERS[1:4])
LG <- linegraph(G)

LLG <- linegraph(LG)

graphreduce(LLG)

if (requireNamespace("Rgraphviz", quietly = TRUE)) {
  plot(graphreduce(LLG))
}

}
