% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ColorList.R
\name{hex12tohex6}
\alias{hex12tohex6}
\title{Convert 12 digit color representations to 6 digit color
  representations}
\usage{
hex12tohex6(x)
}
\details{
Function throws a warning if the conversion looses information. The
  \code{\link{l_hexcolor}} function converts any Tcl color specification to a
  12 digit hexadecimal color representation.
}

