% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_after.R
\name{l_after_idle}
\alias{l_after_idle}
\title{Evaluate a function on once the processor is idle}
\usage{
l_after_idle(fun)
}
\arguments{
\item{fun}{function to be evaluated once tcl interpreter is idle}
}
\description{
It is possible for an observer to call the configure method of 
  that plot while the plot is still in the configuration pipeline. In this 
  case, a warning is thrown as unwanted side effects can happen if the next 
  observer in line gets an outdated notification. In this case, it is 
  recommended to use the l_after_idle function that evaluates some code once 
  the processor is idle.
}
