% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_aspect.R
\name{l_aspect}
\alias{l_aspect}
\title{Query the aspect ratio of a plot}
\usage{
l_aspect(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
aspect ratio
}
\description{
The aspect ratio is defined by the ratio of the number of pixels
  for one data unit on the y axis and the number of pixels for one data unit
  on the x axes.
}
\examples{
p <- with(iris, l_plot(Sepal.Length ~ Sepal.Width, color=Species))

l_aspect(p)
l_aspect(p) <- 1
}
