% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_context_delete}
\alias{l_bind_context_delete}
\title{Delete a context binding}
\usage{
l_bind_context_delete(widget, id)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{id}{context binding id}
}
\description{
Remove a context binding
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_context}}, \code{\link{l_bind_context_ids}},
  \code{\link{l_bind_context_get}}, \code{\link{l_bind_context_reorder}}
}
