% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_context_ids}
\alias{l_bind_context_ids}
\title{List context binding ids}
\usage{
l_bind_context_ids(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
vector with context binding ids
}
\description{
List all user added context binding ids
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_context}}, \code{\link{l_bind_context_get}},
  \code{\link{l_bind_context_delete}}, \code{\link{l_bind_context_reorder}}
}
