% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_glyph_reorder}
\alias{l_bind_glyph_reorder}
\title{Reorder the glyph binding evaluation sequence}
\usage{
l_bind_glyph_reorder(widget, ids)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{ids}{new glyph binding id evaluation order, this must be a 
rearrangement of the elements returned by the
\code{\link{l_bind_glyph_ids}} function.}
}
\value{
vector with binding id evaluation order (same as the \code{id} argument)
}
\description{
The order the glyph bindings defines how they get evaluated 
  once an event matches event patterns of multiple glyph bindings.
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_glyph}}, \code{\link{l_bind_glyph_ids}},
  \code{\link{l_bind_glyph_get}}, \code{\link{l_bind_glyph_delete}}
}
