% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_layer_delete}
\alias{l_bind_layer_delete}
\title{Delete a layer binding}
\usage{
l_bind_layer_delete(widget, id)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{id}{layer binding id}
}
\description{
Remove a layer binding
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_layer}}, \code{\link{l_bind_layer_ids}},
  \code{\link{l_bind_layer_get}}, \code{\link{l_bind_layer_reorder}}
}
