% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_navigator_get}
\alias{l_bind_navigator_get}
\title{Get the event pattern and callback Tcl code of a navigator binding}
\usage{
l_bind_navigator_get(widget, id)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{id}{navigator binding id}
}
\value{
Character vector of length two. First element is the event pattern,
  the second element is the Tcl callback code.
}
\description{
This function returns the registered event pattern and the Tcl
  callback code that the Tcl interpreter evaluates after a event occurs that
  machtches the event pattern.
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_navigator}}, \code{\link{l_bind_navigator_ids}},
  \code{\link{l_bind_navigator_delete}}, \code{\link{l_bind_navigator_reorder}}
}
