% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_navigator_ids}
\alias{l_bind_navigator_ids}
\title{List navigator binding ids}
\usage{
l_bind_navigator_ids(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
vector with navigator binding ids
}
\description{
List all user added navigator binding ids
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_navigator}}, \code{\link{l_bind_navigator_get}},
  \code{\link{l_bind_navigator_delete}}, \code{\link{l_bind_navigator_reorder}}
}
