% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getBinData.R
\name{l_breaks}
\alias{l_breaks}
\title{Gets the boundaries of the histogram bins containing active points.}
\usage{
l_breaks(widget)
}
\arguments{
\item{widget}{A loon histogram widget.}
}
\value{
A named list of the minimum and maximum values of the boundaries for each active bins in the histogram.
}
\description{
Queries the histogram and returns the ids of all active points in each bin that contains active points.
}
\seealso{
\code{\link{l_getBinData}}, \code{\link{l_getBinIds}},
  \code{\link{l_binCut}}
}
