% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_context2d.R
\name{l_context_add_context2d}
\alias{l_context_add_context2d}
\title{Create a context2d navigator context}
\usage{
l_context_add_context2d(navigator, ...)
}
\arguments{
\item{navigator}{navigator handle object}

\item{...}{arguments passed on to modify context states}
}
\value{
context handle
}
\description{
A context2d maps every location on a 2d space graph to a list of
  xvars and a list of yvars such that, while moving the navigator along the
  graph, as few changes as possible take place in xvars and yvars.

Contexts are in more detail explained in the webmanual accessible with 
\code{\link{l_help}}. Please read the section on context by running 
\code{l_help("learn_R_display_graph.html#contexts")}.
}
\seealso{
\code{\link{l_info_states}}, \code{\link{l_context_ids}},
  \code{\link{l_context_add_geodesic2d}},
  \code{\link{l_context_add_slicing2d}}, \code{\link{l_context_getLabel}},
  \code{\link{l_context_relabel}}
}
