% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_context2d.R
\name{l_context_delete}
\alias{l_context_delete}
\title{Delete a context from a navigator}
\usage{
l_context_delete(navigator, id)
}
\arguments{
\item{navigator}{navigator hanlde}

\item{id}{context id}
}
\description{
Navigators can have multiple contexts. This function removes a 
  context from a navigator.
}
\details{
For more information run: \code{l_help("learn_R_display_graph.html#contexts")}
}
\seealso{
\code{\link{l_context_ids}}, \code{\link{l_context_add_context2d}}, 
  \code{\link{l_context_add_geodesic2d}},
  \code{\link{l_context_add_slicing2d}}, \code{\link{l_context_getLabel}},
  \code{\link{l_context_relabel}}
}
