% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_context2d.R
\name{l_context_getLabel}
\alias{l_context_getLabel}
\title{Query the label of a context}
\usage{
l_context_getLabel(navigator, id)
}
\arguments{
\item{navigator}{navigator hanlde}

\item{id}{context id}
}
\description{
Context labels are eventually used in the context inspector. 
  This function queries the label of a context.
}
\details{
For more information run: \code{l_help("learn_R_display_graph.html#contexts")}
}
\seealso{
\code{\link{l_context_getLabel}}, 
  \code{\link{l_context_add_context2d}}, 
  \code{\link{l_context_add_geodesic2d}},
  \code{\link{l_context_add_slicing2d}}, \code{\link{l_context_delete}}
}
