% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_context2d.R
\name{l_context_ids}
\alias{l_context_ids}
\title{List context ids of a navigator}
\usage{
l_context_ids(navigator)
}
\arguments{
\item{navigator}{navigator hanlde}
}
\description{
Navigators can have multiple contexts. This function list the
  context ids of a navigator.
}
\details{
For more information run: \code{l_help("learn_R_display_graph.html#contexts")}
}
\seealso{
\code{\link{l_context_delete}},
  \code{\link{l_context_add_context2d}},
  \code{\link{l_context_add_geodesic2d}},
  \code{\link{l_context_add_slicing2d}}, \code{\link{l_context_getLabel}},
  \code{\link{l_context_relabel}}
}
