% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_compound.R
\name{l_createCompoundGrob}
\alias{l_createCompoundGrob}
\title{For the target compound loon plot, creates the final grob from
the class of the `target`` and the `arrangeGrob.args`}
\usage{
l_createCompoundGrob(target, arrangeGrob.args)
}
\arguments{
\item{target}{the (compound) loon  plot}

\item{arrangeGrob.args}{arguments as described by `gridExtra::arrangeGrob()`}
}
\value{
a grob (or list of grobs) that can be handed to `gTree()` as
`children = gList(returnedValue)` as the final grob constructed for
the compound loon plot.  Default for an `l_compound` is to simply execute
`gridExtra::arrangeGrob(arrangeGrob.args)`.
}
\description{
For the target compound loon plot, creates the final grob from
the class of the `target`` and the `arrangeGrob.args`
}
