% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_currenttags.R
\name{l_currenttags}
\alias{l_currenttags}
\title{Get tags of the item below the mouse cursor}
\usage{
l_currenttags(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
vector with item tags of visual
}
\description{
Retrieves the tags of the visual item that at the time of the
  function evaluation is below the mouse cursor.
}
\details{
For more details see \code{l_help("learn_R_bind.html#item-bindings")}
}
\examples{
printTags <- function(W) {
    print(l_currenttags(W))
}

p <- l_plot(x=1:3, y=1:3, title='Query Visual Item Tags')

l_bind_item(p, 'all', '<ButtonPress>', function(W)printTags(W))
}
\seealso{
\code{\link{l_bind_item}}, \code{\link{l_currentindex}}
}
