% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_data.R
\name{l_data}
\alias{l_data}
\title{Convert an R data.frame to a Tcl dictionary}
\usage{
l_data(data)
}
\arguments{
\item{data}{a data.frame object}
}
\value{
a string that represents with data.frame with a Tcl dictionary data
  structure.
}
\description{
This is a helper function to convert an \R data.frame object to 
  a Tcl data frame object. This function is useful when changing a data state
  with \code{\link{l_configure}}.
}
