% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_export.R
\name{l_export_valid_formats}
\alias{l_export_valid_formats}
\title{Return a list of the available image formats when exporting a loon 
  plot}
\usage{
l_export_valid_formats()
}
\value{
a vector with the image formats available for exporting a loon plot.
}
\description{
The supported image formats are dependent on the system 
  environment. Plots can always be exported to the Postscript format. 
  Exporting displays as .pdfs is only possible when the command line tool 
  epstopdf is installed. Finally, exporting to either png, jpg, bmp, tiff or 
  gif requires the Img Tcl extension. When choosing one of the formats that 
  depend on the Img extension, it is possible to export any Tk widget as an 
  image including inspectors.
}
