% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_facet_grid.R
\name{l_facet_grid}
\alias{l_facet_grid}
\title{Layout Facets in a grid}
\usage{
l_facet_grid(
  widget,
  by,
  linkingGroup,
  inheritLayers = TRUE,
  byrow = TRUE,
  column_labels_loc = c("top", "bottom"),
  row_labels_loc = c("right", "left"),
  span = 10,
  label_background = "gray80",
  label_foreground = "black",
  label_borderwidth = 2,
  label_relief = c("groove", "flat", "raised", "sunken", "ridge", "solid"),
  parent = NULL,
  ...
)
}
\arguments{
\item{widget}{A loon widget}

\item{by}{Faceting categories. It could be "color", "size", etc.
Available faceting categories can be achieved by \code{l_nDimStateNames(your widget)}}

\item{linkingGroup}{A linkingGroup for widgets. If missing, default would be a paste of
"facet" and the current tk path number}

\item{inheritLayers}{Logical. Should widget layers be inherited into faceting panels?}

\item{byrow}{Are labels display by row or not?}

\item{column_labels_loc}{Column labels location, one of "top" or "bottom".}

\item{row_labels_loc}{Row labels location, one of "right" or "left".}

\item{span}{It is used to control the proportion of panel size (height or width) and
the label size (label height and label width)}

\item{label_background}{Label background color}

\item{label_foreground}{Label foreground color}

\item{label_borderwidth}{Label border width}

\item{label_relief}{Label relief}

\item{parent}{parent widget path}

\item{...}{named arguments to modify the `loon` widget states}
}
\value{
an `l_facet` object (an `l_compound` object), being a list with named elements,
each representing a separate interactive plot.
The names of the plots should be self explanatory and a list
of all plots can be accessed from the `l_facet` object via `l_getPlots()`.
}
\description{
It takes a loon widget and forms a matrix of loon widgets
faceting by aesthetics ("color", "size", "glyph", etc). It can handle mutiple (more than 2) faceting categories
}
\examples{
library(maps)
p <- with(quakes, l_plot(long, lat, linkingGroup = "quakes"))
p["color"][quakes$mag < 5 & quakes$mag >= 4] <- "lightgreen"
p["color"][quakes$mag < 6 & quakes$mag >= 5] <- "lightblue"
p["color"][quakes$mag >= 6] <- "firebrick"
# A Fiji map
NZFijiMap <- map("world2", regions = c("New Zealand", "Fiji"), plot = FALSE)
l_layer(p, NZFijiMap,
        label = "New Zealand and Fiji",
        color = "forestgreen",
        index = "end")
fp <- l_facet_grid(p, by = "color", linkingGroup = "quakes")

s <- l_serialaxes(iris, color = iris$Species, scaling = "observation")
fs <- l_facet_grid(s, by = "color")
# The linkingGroup can be printed or accessed by
l_configure(s, linkingGroup = fs[[1]]['linkingGroup'], sync = "pull")
}
