% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_facet_wrap.R
\name{l_facet_wrap}
\alias{l_facet_wrap}
\title{Layout Facets in a grid}
\usage{
l_facet_wrap(
  widget,
  by,
  linkingGroup,
  inheritLayers = TRUE,
  scales = c("fixed", "fixed_x", "fixed_y", "free_x", "free_y", "free"),
<<<<<<< HEAD
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  labels_loc = c("top", "bottom"),
  span = 10,
  label_background = "gray80",
  label_foreground = "black",
  label_borderwidth = 2,
  label_relief = c("groove", "flat", "raised", "sunken", "ridge", "solid"),
  parent = NULL,
  ...
)
=======
  nrow = NULL, ncol = NULL, byrow = TRUE, labels_loc = c("top",
  "bottom"), span = 10, label_background = "gray80",
  label_foreground = "black", label_borderwidth = 2,
  label_relief = c("groove", "flat", "raised", "sunken", "ridge",
  "solid"), parent = NULL, ...)
>>>>>>> upstream/master
}
\arguments{
\item{widget}{A loon widget}

\item{by}{Faceting categories. It could be "color", "size", etc.
Available faceting categories can be achieved by \code{l_nDimStateNames(your widget)}}

\item{linkingGroup}{A linkingGroup for widgets. If missing, default would be a paste of
"facet" and the current tk path number}

\item{inheritLayers}{Logical. Should widget layers be inherited into faceting panels?}

\item{scales}{All panels can share the same scales if it is "fixed"; or all panels only vary the
scales across rows ("free_x" or "fixed_y"); or all panels only vary the
scales across columns ("free_y" or "fixed_x"); or their scales are totally free ("free"). Note that when rows or
columns share the same scale, the dynamic chages of scales will be synchronized.}

\item{nrow}{The number of layout rows}

\item{ncol}{The number of layout columns}

\item{byrow}{Are labels display by row or not?}

\item{labels_loc}{Labels location, "top" or "bottom"?}

\item{span}{It is used to control the proportion of panel size (height or width) and
the label size (label height and label width)}

\item{label_background}{Label background color}

\item{label_foreground}{Label foreground color}

\item{label_borderwidth}{Label border width}

\item{label_relief}{Label relief}

\item{parent}{parent widget path}

\item{...}{named arguments to modify the `loon` widget states}
}
\value{
an `l_facet` object (an `l_compound` object), being a list with named elements,
each representing a separate interactive plot.
The names of the plots should be self explanatory and a list
of all plots can be accessed from the `l_facet` object via `l_getPlots()`.
}
\description{
It takes a loon widget and forms a matrix of loon widgets
faceting by aesthetics ("color", "size", "glyph", etc). It can handle mutiple (more than 2) faceting categories
}
\examples{
p <- with(mtcars, l_plot3D(mpg, hp, wt, color = cyl))
fp <- l_facet_wrap(p, by = "color")

}
