% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_compound.R
\name{l_get_arrangeGrobArgs}
\alias{l_get_arrangeGrobArgs}
\title{For the target (compound) loon plot, determines all arguments
(i.e. including the grobs) to be passed to `gridExtra::arrangeGrob()`
so as to determine the layout in `grid` graphics.}
\usage{
l_get_arrangeGrobArgs(target)
}
\arguments{
\item{target}{the (compound) loon  plot to be laid out.}
}
\value{
a list of the named arguments and their values to be passed to
`gridExtra::arrangeGrob()`.
}
\description{
For the target (compound) loon plot, determines all arguments
(i.e. including the grobs) to be passed to `gridExtra::arrangeGrob()`
so as to determine the layout in `grid` graphics.
}
