% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getBinData.R
\name{l_getBinData}
\alias{l_getBinData}
\title{Get information on current bins from a histogram}
\usage{
l_getBinData(widget)
}
\arguments{
\item{widget}{A loon histogram widget.}
}
\value{
A nested list of the bins in the histogram which contain active points.
Each bin is a list of the counts, the point indices, and the minimum (x0) and maximum (x1)
of that bin.  Loon histogram bins are open  on the left and closed on the right by default, namely "(x0, x1]".
The counts and the points further identify the number and ids of all points,
those which are selected, and those of each colour in that bin (identified by their hex12 colour from tcl).
}
\description{
Queries the histogram and returns information about all active cases
contained by the histogram's bins.
}
\seealso{
\code{\link{l_getBinIds}}, \code{\link{l_breaks}},
  \code{\link{l_binCut}}
}
