% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getBinData.R
\name{l_getBinIds}
\alias{l_getBinIds}
\title{Gets the ids of the active points in each bin of a histogram}
\usage{
l_getBinIds(widget)
}
\arguments{
\item{widget}{A loon histogram widget.}
}
\value{
A named list of the bins in the histogram and the ids of their active points.
}
\description{
Queries the histogram and returns the ids of all active points in each bin that contains active points.
}
\seealso{
\code{\link{l_getBinData}}, \code{\link{l_breaks}},
  \code{\link{l_binCut}}
}
