% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ColorList.R
\name{l_getColorList}
\alias{l_getColorList}
\title{Get loon's color mapping list}
\usage{
l_getColorList()
}
\value{
a vector with hex-encoded colors
}
\description{
The color mapping list is used by loon to convert nominal values
  to color values, see the documentation for \code{\link{l_setColorList}}.
}
\seealso{
\code{\link{l_setColorList}}
}
