% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_linkingWarning.R
\name{l_getDeprecatedLinkedVar}
\alias{l_getDeprecatedLinkedVar}
\title{A helper function to determine which, if any, linked variables in a plot are now deprecated.}
\usage{
l_getDeprecatedLinkedVar(
  widget,
  args,
  modifiedLinkedStates = character(0L),
  l_className = NULL
)
}
\arguments{
\item{widget}{the loon widget or a list of loon widgets}

\item{args}{named list of state values for linkable state variables named in the list.}

\item{modifiedLinkedStates}{vector of variable names having modified link states/}

\item{l_className}{class of the loon widget}
}
\value{
a named vector, or list of vectors, of logicals indicating whether these linkable states need be deprecated or not.
}
\description{
Checks which linkable states are linked, whether their input state is the default one, valid,
and equal to the plot current states.  It is not intended for the general user.
}
\keyword{internal}
