% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getFromPath.R
\name{l_getFromPath}
\alias{l_getFromPath}
\title{Create loon objects from path name}
\usage{
l_getFromPath(target)
}
\arguments{
\item{target}{loon object specification (e.g. \code{".l0.plot"})}
}
\description{
This function can be used to create the loon objects from
  a valid widget path name. The main difference from \code{l_create_handle} is that
  \code{l_getFromPath} can take a loon compound widget path
  but \code{l_create_handle} cannot.
}
\details{
For more information run: \code{l_help("learn_R_intro.html#re-creating-object-handles")}
}
\examples{
\dontrun{
 l_pairs(iris, showHistogram = TRUE)
 # The path can be found at the top of tk title
 # Suppose it is the first loon widget, this path should be ".l0.pairs"
 p <- l_create_handle(".l0.pairs") # error
 p <- l_getFromPath(".l0.pairs")
}
}
\seealso{
\code{\link{l_create_handle}} \code{\link{l_loonWidgets}}
}
