% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getGraph.R
\name{l_getGraph}
\alias{l_getGraph}
\title{Extract a loongraph or graph object from loon's graph display}
\usage{
l_getGraph(widget, asloongraph = TRUE)
}
\arguments{
\item{widget}{a graph widget handle}

\item{asloongraph}{boolean, if TRUE then the function returns a loongraph
object, otherwise the function returns a graph object defined in the graph
\R package.}
}
\value{
a loongraph or a graph object
}
\description{
The graph display represents a graph with the \code{nodes}, 
  \code{from}, \code{to}, and \code{isDirected} plot states. This function 
  creates a loongraph or a graph object using these states.
}
\seealso{
\code{\link{l_graph}}, \code{\link{loongraph}}
}
