% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_linkedStates.R
\name{l_getLinkedStates}
\alias{l_getLinkedStates}
\title{Query the States that are Linked with Loon's Standard Linking Model}
\usage{
l_getLinkedStates(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
vector with state names that are linked states
}
\description{
Loon's standard linking model is based on three levels, the 
  \code{linkingGroup} and \code{linkingKey} states and the \emph{used 
  linkable states}. See the details in the documentation for
  \code{\link{l_setLinkedStates}}.
}
\seealso{
\code{\link{l_setLinkedStates}}
}
