% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getOption.R
\name{l_getOption}
\alias{l_getOption}
\title{Get the value of a loon display option}
\usage{
l_getOption(option)
}
\arguments{
\item{option}{the name of the option being queried.}
}
\value{
the value of the named option.
}
\description{
All of loon's displays access a set of common options.  
This function accesses and returns the current value of the named option.
}
\examples{
l_getOption("background")

}
\seealso{
\code{\link{l_getOptionNames}}, \code{\link{l_userOptions}},
  \code{\link{l_userOptionDefault}}, \code{\link{l_setOption}}
}
