% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getOption.R
\name{l_getOptionNames}
\alias{l_getOptionNames}
\title{Get the names of all loon display options}
\usage{
l_getOptionNames()
}
\value{
a vector of all loon display option names.
}
\description{
All of loon's displays access a set of common options.
This function accesses and returns the names of all loon options.
}
\examples{
l_getOptionNames()

}
\seealso{
\code{\link{l_getOption}}, \code{\link{l_userOptions}},
  \code{\link{l_userOptionDefault}}, \code{\link{l_setOption}}
}
