% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_compound.R, R/l_facet.R, R/l_pairs.R,
%   R/l_plot_ts.R
\name{l_getPlots}
\alias{l_getPlots}
\alias{l_getPlots.l_facet}
\alias{l_getPlots.l_pairs}
\alias{l_getPlots.l_ts}
\title{For the target compound loon plot, determines all the loon plots
in that compound plot.}
\usage{
l_getPlots(target)

\method{l_getPlots}{l_facet}(target)

\method{l_getPlots}{l_pairs}(target)

\method{l_getPlots}{l_ts}(target)
}
\arguments{
\item{target}{the (compound) loon  plot to be laid out.}
}
\value{
a list of the named arguments and their values to be passed to
`gridExtra::arrangeGrob()`.
}
\description{
For the target compound loon plot, determines all the loon plots
in that compound plot.
}
