% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getScaledData.R
\name{l_getScaledData}
\alias{l_getScaledData}
\title{scale data}
\usage{
l_getScaledData(
  data,
  sequence = NULL,
  scaling = c("variable", "observation", "data", "none"),
  displayOrder = NULL,
  reserve = FALSE,
  as.data.frame = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{sequence}{vector with variable names that defines the axes sequence.
If \code{NULL}, it will be set as the column names automatically.}

\item{scaling}{one of 'variable', 'data', 'observation' or 'none' to specify how the data is scaled.}

\item{displayOrder}{the order of the display}

\item{reserve}{If \code{TRUE}, return the variables not shown in \code{sequence} as well;
else only return the variables defined in \code{sequence}.}

\item{as.data.frame}{Return a matrix or a data.frame}
}
\description{
It is mainly used in serial axes
}
