% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_glyphs.R
\name{l_glyph_add.default}
\alias{l_glyph_add.default}
\title{Default method for adding non-primitive and loon-native glyphs}
\usage{
\method{l_glyph_add}{default}(widget, type, ...)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{type}{loon-native non-primitive glyph type, one of \code{'text'}, 
\code{'serialaxes'}, \code{'image'}, \code{'[polygon'}, or
\code{'pointrange'}}

\item{...}{state arguments}
}
\description{

}

