% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{l_glyphs_inspector_pointrange}
\alias{l_glyphs_inspector_pointrange}
\title{Create a Pointrange Glyph Inspector}
\usage{
l_glyphs_inspector_pointrange(parent = NULL, ...)
}
\arguments{
\item{parent}{parent widget path}

\item{...}{state arguments}
}
\value{
widget handle
}
\description{
Inpectors provide graphical user interfaces to oversee and
  modify plot states
}
\examples{
if(interactive()){

i <- l_glyphs_inspector_pointrange()
}
}
\seealso{
\code{\link{l_create_handle}}
}
