% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graph.R
\name{l_graph.graph}
\alias{l_graph.graph}
\title{Create a graph display based on a graph object}
\usage{
\method{l_graph}{graph}(nodes, ...)
}
\arguments{
\item{nodes}{a graph object created with the functions in the \code{graph} \R
package.}

\item{...}{arguments to modify the graph display state}
}
\value{
graph handle
}
\description{
Graph objects are defined in the graph \R package.
}
\details{
For more information run: \code{l_help("learn_R_display_graph.html#graph")}
}
\seealso{
\code{\link{l_graph}}, \code{\link{l_info_states}},
  \code{\link{l_graph.loongraph}}
}
