% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graph.R
\name{l_graph.loongraph}
\alias{l_graph.loongraph}
\title{Create a graph display based on a loongraph object}
\usage{
\method{l_graph}{loongraph}(nodes, ...)
}
\arguments{
\item{nodes}{a loongraph object created with the \code{\link{loongraph}}
function.}

\item{...}{arguments to modify the graph display state}
}
\value{
graph handle
}
\description{
Loongraphs can be created with the \code{\link{loongraph}}
  function.
}
\details{
For more information run: \code{l_help("learn_R_display_graph.html#graph")}
}
\seealso{
\code{\link{loongraph}}, \code{\link{l_graph}},
  \code{\link{l_info_states}}, \code{\link{l_graph.graph}}
}
