% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{l_graph_inspector_navigators}
\alias{l_graph_inspector_navigators}
\title{Create a Graph Navigator Inspector}
\usage{
l_graph_inspector_navigators(parent = NULL, ...)
}
\arguments{
\item{parent}{parent widget path}

\item{...}{state arguments}
}
\value{
widget handle
}
\description{
Inpectors provide graphical user interfaces to oversee and
  modify plot states
}
\examples{
i <- l_graph_inspector_navigators()
}
\seealso{
\code{\link{l_create_handle}}
}
