% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_add.default}
\alias{l_graphswitch_add.default}
\title{Add a graph that is defined by node names and a from-to edges list}
\usage{
\method{l_graphswitch_add}{default}(widget, graph, from, to, isDirected,
  label = "", index = "end", ...)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{graph}{a vector with the node names, i.e. this argument gets passed on
as the nodes argument to creat a \code{\link{loongraph}} like object}

\item{from}{vector with node names of the from-to pairs for edges}

\item{to}{vector with node names of the from-to pairs for edges}

\item{isDirected}{boolean to indicate whether the from-to-list defines 
directed or undirected edges}

\item{label}{string with label for graph}

\item{index}{position of graph in the graph list}

\item{...}{additional arguments are not used for this methiod}
}
\value{
id for graph in the graphswitch widget
}
\description{
This default method uses the loongraph display states as 
  arguments to add a graph to the graphswitch widget.
}
\seealso{
\code{\link{l_graphswitch}}
}
