% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_add.graph}
\alias{l_graphswitch_add.graph}
\title{Add a graph to the graphswitch widget using a graph object}
\usage{
\method{l_graphswitch_add}{graph}(widget, graph, label = "", index = "end", ...)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{graph}{a graph object created with the functions in the \code{graph} \R
package.}

\item{label}{string with label for graph}

\item{index}{position of graph in the graph list}

\item{...}{additional arguments are not used for this methiod}
}
\value{
id for graph in the graphswitch widget
}
\description{
Graph objects are defined in the graph \R package.
}
\seealso{
\code{\link{l_graphswitch}}
}
