% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_add.loongraph}
\alias{l_graphswitch_add.loongraph}
\title{Add a graph to the graphswitch widget using a loongraph object}
\usage{
\method{l_graphswitch_add}{loongraph}(widget, graph, label = "", index = "end", ...)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{graph}{a loongraph object}

\item{label}{string with label for graph}

\item{index}{position of graph in the graph list}

\item{...}{additional arguments are not used for this methiod}
}
\value{
id for graph in the graphswitch widget
}
\description{
Loongraphs can be created with the \code{\link{loongraph}}
  function.
}
\seealso{
\code{\link{l_graphswitch}}
}
