% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_add}
\alias{l_graphswitch_add}
\title{Add a graph to a graphswitch widget}
\usage{
l_graphswitch_add(widget, graph, ...)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{graph}{a graph or a loongraph object}

\item{...}{arguments passed on to method}
}
\value{
id for graph in the graphswitch widget
}
\description{
This is a generic function to add a graph to a graphswitch
  widget.
}
\details{
For more information run: \code{l_help("learn_R_display_graph.html#graph-switch-widget")}
}
\seealso{
\code{\link{l_graphswitch}}
}
