% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_delete}
\alias{l_graphswitch_delete}
\title{Delete a graph from the graphswitch widget}
\usage{
l_graphswitch_delete(widget, id)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{id}{of the graph}
}
\description{
Remove a a graph from the graphswitch widget
}
\seealso{
\code{\link{l_graphswitch}}
}
