% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_getLabel}
\alias{l_graphswitch_getLabel}
\title{Query Label of a Graph in the Graphswitch Widget}
\usage{
l_graphswitch_getLabel(widget, id)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{id}{of the graph}
}
\description{
The graphs in the graphswitch widgets have labels. Use this
  function to query the label of a graph.
}
\seealso{
\code{\link{l_graphswitch}}
}
