% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_ids}
\alias{l_graphswitch_ids}
\title{List the ids of the graphs in the graphswitch widget}
\usage{
l_graphswitch_ids(widget)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}
}
\description{
Every graph in the graphswitch widget has an id. This function
  returns these ids preserving the oder of how the graphs are listed in the
  graphswitch.
}
