% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_move}
\alias{l_graphswitch_move}
\title{Move a Graph in the Graph List}
\usage{
l_graphswitch_move(widget, id, index)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{id}{of the graph}

\item{index}{position of the graph as a positive integer, \code{"start"} and
\code{"end"} are also valid keywords.}
}
\description{
Change the postion in of a graph in the graphswitch widget.
}
\seealso{
\code{\link{l_graphswitch}}
}
