% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_relabel}
\alias{l_graphswitch_relabel}
\title{Relabel a Graph in the Graphswitch Widget}
\usage{
l_graphswitch_relabel(widget, id, label)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{id}{of the graph}

\item{label}{string with label of graph}
}
\description{
The graphs in the graphswitch widgets have labels. Use this
  function the relabel a graph.
}
\seealso{
\code{\link{l_graphswitch}}
}
