% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_reorder}
\alias{l_graphswitch_reorder}
\title{Reorder the Positions of the Graphs in the Graph List}
\usage{
l_graphswitch_reorder(widget, ids)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{ids}{vector with all graph ids from the graph widget. Use
\code{\link{l_graphswitch_ids}} to query the ids.}
}
\description{
Define a new graph order in the graph list.
}
\seealso{
\code{\link{l_graphswitch}}
}
