% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_set}
\alias{l_graphswitch_set}
\title{Change the Graph shown in the Active Graph Widget}
\usage{
l_graphswitch_set(widget, id)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{id}{of the graph}
}
\description{
The \code{activewidget} state holds the widget handle of a graph
  display. This function replaces the graph in the \code{activewidget} with
  one of the graphs in the graphswitch widget.
}
\seealso{
\code{\link{l_graphswitch}}
}
