% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_help.R
\name{l_help}
\alias{l_help}
\title{Open a browser with loon's local documentation webpage}
\usage{
l_help(page = "index", ...)
}
\arguments{
\item{page}{relative path to a page, the .html part may be omitted}

\item{...}{arguments forwarded to browseURL, e.g. to specify a browser}
}
\description{
The loon R package comes with a copy of loon's official webpage 
  at \url{http://waddella.github.io/loon/}. \code{l_help} opens a browser 
  with this page.
}
\examples{

\\donotrun{
l_help()
l_help("learn_R_display_hist")
l_help("learn_R_bind")
# jump to a section
l_help("learn_R_bind.html#list-reorder-delete-bindings")
}
}

