% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_hexcolor.R
\name{l_hexcolor}
\alias{l_hexcolor}
\title{Convert color names to their 12 digit hexadecimal color representation}
\usage{
l_hexcolor(colors)
}
\description{
Color names in loon will be mapped to colors according to the Tk
  color specifications and are normalized to a 12 digit hexadecimal color 
  representation. This function maps Tk color names to 
  
  
  Colors in color states in loon are normalized usto a 12 digit hexadecimal 
  color representatio
}
\examples{
p <- l_plot(1:2)
p['color'] <- 'red'
p['color']

l_hexcolor('red')
}

