% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_hist.R
\name{l_hist}
\alias{l_hist}
\title{Create an interactive histogram}
\usage{
l_hist(x, ...)
}
\arguments{
\item{x}{vector with numerical data to perform the binning on x,}

\item{...}{named arguments to modify the histogram plot states or layouts, see details.}
}
\value{
widget handle
}
\description{
\code{l_hist} is a generic function for creating interactive visualization environments for R objects.
}
\details{
For more information run: \code{l_help("learn_R_display_hist")}
}
\examples{
if(interactive()) {
  # numerical data
  x <- rnorm(100)
  h <- l_hist(x)

  # categorical data
  x <- factor(mtcars$cyl)
  h <- l_hist(x)
}
}
\seealso{
\code{\link{l_hist.default}}, \code{\link{l_plot}}
}
