% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_image_import_files.R
\name{l_image_import_files}
\alias{l_image_import_files}
\title{Import Image Files as Tk Image Objects}
\usage{
l_image_import_files(paths)
}
\arguments{
\item{paths}{vector with paths to image files that are supported}
}
\value{
vector of image object names
}
\description{
Note that the supported image file formats depend on whether the
  \code{Img} Tk extension is installed.
}
\details{
For more information run: \code{l_help("learn_R_display_plot.html#load-images")}
}
\seealso{
\code{\link{l_image_import_array}}, \code{\link{l_imageviewer}}
}
