% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_isLoonWidget.R
\name{l_isLoonWidget}
\alias{l_isLoonWidget}
\title{Check if a widget path is a valid loon widget}
\usage{
l_isLoonWidget(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
boolean, TRUE if the argument is a valid loon widget path, FALSE
  otherwise
}
\description{
This function can be useful to check whether a loon widget is
  has been closed by the user.
}
