% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer.R
\name{l_layer_getParent}
\alias{l_layer_getParent}
\title{Get parent layer id of a layer}
\usage{
l_layer_getParent(widget, layer)
}
\arguments{
\item{widget}{widget path or layer object of class \code{'l_layer'}}

\item{layer}{layer id. If the widget argument is of class \code{'l_layer'}
then the layer argument is not used}
}
\description{
The toplevel parent is the \code{'root'} layer.
}
\examples{
p <- with(iris, l_plot(Sepal.Length ~ Sepal.Width, color=Species))

l_layer_getParent(p, 'model')
}
\seealso{
\code{\link{l_layer}}, \code{\link{l_layer_getChildren}}
}
