% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer.R
\name{l_layer_getType}
\alias{l_layer_getType}
\title{Get layer type}
\usage{
l_layer_getType(widget, layer)
}
\arguments{
\item{widget}{widget path or layer object of class \code{'l_layer'}}

\item{layer}{layer id. If the widget argument is of class \code{'l_layer'}
then the layer argument is not used}
}
\value{
One of: \code{'group'}, \code{'polygon'}, \code{'text'},
  \code{'line'}, \code{'rectangle'}, \code{'oval'}, \code{'points'},
  \code{'texts'}, \code{'polygons'}, \code{'rectangles'}, \code{'lines'} and
  \code{'scatterplot'}, \code{'histogram'}, \code{'serialaxes'} and
  \code{'graph'}.
}
\description{
To see the manual page of \code{\link{l_layer}} for all the
  primitive layer types.
}
\details{
For more information run: \code{l_help("learn_R_layer")}
}
\examples{
if(interactive()){

p <- l_plot()
l <- l_layer_rectangle(p, x=0:1, y=0:1)
l_layer_getType(p, l)
l_layer_getType(p, 'model')

}
}
\seealso{
\code{\link{l_layer}}
}
