% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer.R
\name{l_layer_index}
\alias{l_layer_index}
\title{Get the order index of a layer among its siblings}
\usage{
l_layer_index(widget, layer)
}
\arguments{
\item{widget}{widget path or layer object of class \code{'l_layer'}}

\item{layer}{layer id. If the widget argument is of class \code{'l_layer'}
then the layer argument is not used}
}
\value{
numeric value
}
\description{
The index determines the rendering order of the children layers
  of a parent. The layer with index=0 is rendered first.
}
\details{
Note that the index for layers is 0 based.
}
\seealso{
\code{\link{l_layer}}, \code{\link{l_layer_move}}
}
