% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer.R
\name{l_layer_points}
\alias{l_layer_points}
\title{Layer a points}
\usage{
l_layer_points(widget, x, y = NULL, color = "gray60", size = 6,
  label = "points", parent = "root", index = 0, ...)
}
\arguments{
\item{widget}{widget path name as a string}

\item{x}{the coordinates of line. Alternatively, a single plotting structure,
function or any \emph{R} object with a plot method can be provided as x and
y are passed on to \code{\link{xy.coords}}}

\item{y}{the y coordinates of the line, optional if x is an appropriate 
structure.}

\item{color}{color of points}

\item{size}{size point, as for scatterplot model layer}

\item{label}{label used in the layers inspector}

\item{parent}{group layer}

\item{index}{of the newly added layer in its parent group}

\item{...}{additional state initialization arguments, see 
\code{\link{l_info_states}}}
}
\value{
layer object handle, layer id
}
\description{
Loon's displays that are based on Cartesian coordinates (i.e.
  scatterplot, histogram and graph display) allow for layering visual
  information including polygons, text and rectangles.

Scatter points layer
}
\details{
For more information run: \code{l_help("learn_R_layer")}
}
\seealso{
\code{\link{l_layer}}, \code{\link{l_info_states}}
}
