% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer.R
\name{l_layer_printTree}
\alias{l_layer_printTree}
\title{Print the layer tree}
\usage{
l_layer_printTree(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
empty string
}
\description{
Prints the layer tree (i.e. the layer ids) to the prompt. Group
  layers are prefixed with a \code{'+'}. The \code{'root'} layer is not
  listed.
}
\examples{
if(interactive()){

p <- l_plot()
l_layer_rectangle(p, x=0:1, y=0:1)
g <- l_layer_group(p)
l_layer_oval(p, x=0:1, y=0:1, parent=g)
l_layer_line(p, x=0:1, y=0:1, parent=g)
l_layer_printTree(p)

}
}
\seealso{
\code{\link{l_layer}}, \code{\link{l_layer_getChildren}}, \code{\link{l_layer_getParent}}
}
