% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer.R
\name{l_layer_relabel}
\alias{l_layer_relabel}
\title{Change layer label}
\usage{
l_layer_relabel(widget, layer, label)
}
\arguments{
\item{widget}{widget path or layer object of class \code{'l_layer'}}

\item{layer}{layer id. If the widget argument is of class \code{'l_layer'}
then the layer argument is not used}

\item{label}{new label of layer}
}
\value{
0 if success otherwise the function throws an error
}
\description{
Layer labels are useful to identify layer in the layer 
  inspector. The layer label can be initially set at layer creation with the
  label argument.
}
\details{
Note that the layer label is not a state of the layer itself, 
  instead is information that is part of the layer collection (i.e. its 
  parent widget).
}
\examples{
if(interactive()){

p <- l_plot()

l <- l_layer_rectangle(p, x=0:1, y=0:1, label="A rectangle")
l_layer_getLabel(p, l)

l_layer_relabel(p, l, label="A relabelled rectangle")
l_layer_getLabel(p, l)

}
}
\seealso{
\code{\link{l_layer}}, \code{\link{l_layer_getLabel}}
}
