% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{l_loon_inspector}
\alias{l_loon_inspector}
\title{Create a loon inspector}
\usage{
l_loon_inspector(parent = NULL, ...)
}
\arguments{
\item{parent}{a valid Tk parent widget path. When the parent widget is
specified (i.e. not \code{NULL}) then the plot widget needs to be placed using
some geometry manager like \code{\link{tkpack}} or \code{\link{tkplace}} in
order to be displayed. See the examples below.}

\item{...}{state arguments, see \code{\link{l_info_states}}.}
}
\value{
a loon widget
}
\description{
The loon inspector is a singleton widget that provids an
  overview to view and modify the active plot.
}
\details{
For more information run: \code{l_help("learn_R_display_inspectors")}
}
\examples{
if(interactive()){
i <- l_loon_inspector()
}
}
