% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navgraph.R
\name{l_navgraph}
\alias{l_navgraph}
\title{Create a Navgraph}
\usage{
l_navgraph(data, separator = ":", graph = NULL, ...)
}
\description{
Create a Navgraph
}
\examples{
ng <- l_navgraph(oliveAcids, color=olive$Area)
ng2 <- l_navgraph(oliveAcids, separator='-', color=olive$Area)
}

