% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navigator.R
\name{l_navigator_delete}
\alias{l_navigator_delete}
\title{Delete a Navigator}
\usage{
l_navigator_delete(widget, id)
}
\arguments{
\item{widget}{graph widget}

\item{id}{navigator handle or navigator id}
}
\description{
Removes a navigator from a graph widget
}
\seealso{
\code{\link{l_navigator_add}}
}
