% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navigator.R
\name{l_navigator_getPath}
\alias{l_navigator_getPath}
\title{Get the sequence of nodes of a navigator's current path}
\usage{
l_navigator_getPath(navigator)
}
\arguments{
\item{navigator}{navigator handle}
}
\value{
a vector of node names for the current path of the navigator
}
\description{
Determines and returns the current path of the navigator.
}
