% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navigator.R
\name{l_navigator_ids}
\alias{l_navigator_ids}
\title{List Navigators}
\usage{
l_navigator_ids(widget)
}
\arguments{
\item{widget}{graph widget}
}
\description{
Lists all navigators that belong to a graph
}
\seealso{
\code{\link{l_navigator_add}}
}
