% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navigator.R
\name{l_navigator_relabel}
\alias{l_navigator_relabel}
\title{Modify the Label of a Navigator}
\usage{
l_navigator_relabel(widget, id, label)
}
\arguments{
\item{widget}{graph widget handle}

\item{id}{navigator id}

\item{label}{new label of navigator}
}
\description{
Change the navigator label
}
\seealso{
\code{\link{l_navigator_add}}
}
