% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navigator.R
\name{l_navigator_walk_backward}
\alias{l_navigator_walk_backward}
\title{Have the Navigator Walk Backward on the Current Path}
\usage{
l_navigator_walk_backward(navigator, to = "")
}
\arguments{
\item{navigator}{navigator handle}

\item{to}{node name that is part of the active path backward where the 
navigator should stop.}
}
\description{
Animate a navigator by having it walk on a path on the graph
}
\details{
Note that navigators have the states \code{animationPause} and 
  \code{animationProportionIncrement} to control the animation speed. 
  Further, you can stop the animation when clicking somewhere on the graph 
  display or by using the mouse scroll wheel.
}
\seealso{
\code{\link{l_navigator_add}}
}
