% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_navigator.R
\name{l_navigator_walk_path}
\alias{l_navigator_walk_path}
\title{Have the Navigator Walk a Path on the Graph}
\usage{
l_navigator_walk_path(navigator, path)
}
\arguments{
\item{navigator}{navigator handle}

\item{path}{vector with node names of the host graph that form a valid path
on that graph}
}
\description{
Animate a navigator by having it walk on a path on the graph
}
\seealso{
\code{\link{l_navigator_add}}
}
