% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_nDimStateNames.R
\name{l_nDimStateNames}
\alias{l_nDimStateNames}
\title{N dimensional state names access}
\usage{
l_nDimStateNames(loon_plot)
}
\arguments{
\item{loon_plot}{A \code{loon} widget or the class name of a loon plot}
}
\description{
Get all n dimensional state names
}
\examples{
if(interactive()){

p <- l_plot()
l_nDimStateNames(p)
l_nDimStateNames("l_plot")

}
}
