% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedLists.R
\name{l_nestedTclList2Rlist}
\alias{l_nestedTclList2Rlist}
\title{Convert a Nested Tcl List to an R List}
\usage{
l_nestedTclList2Rlist(tclobj, transform = function(x) {
    as.numeric(x)
})
}
\arguments{
\item{tclobj}{a tcl object as returned by \code{\link[tcltk:TkCommands]{tcl}} or
\code{\link[tcltk:TclInterface]{.Tcl}}.}

\item{transform}{a function to transfrom the string output to another data
type}
}
\value{
a nested \R list
}
\description{
Helper function to work with \R and Tcl
}
\examples{

tclobj <- .Tcl('set a {{1 2 3} {2 3 4 4} {3 5 3 3}}')
l_nestedTclList2Rlist(tclobj)
}
\seealso{
\code{\link{l_Rlist2nestedTclList}}
}
