% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ng_plots.R
\name{l_ng_plots.measures}
\alias{l_ng_plots.measures}
\title{Select 2d spaces in scatterplot matrix}
\usage{
\method{l_ng_plots}{measures}(measures, ...)
}
\description{
Select 2d spaces in scatterplot matrix
}
\examples{
# 2d measures
scags <- scagnostics2d(oliveAcids, separator='**')
scags()
ng <- l_ng_plots(scags, color=olive$Area)

# 1d measures
scale01 <- function(x){(x-min(x))/diff(range(x))}
m1d <- measures1d(sapply(iris[,-5], scale01),
     mean=mean, median=median, sd=sd,
     q1=function(x)as.vector(quantile(x, probs=0.25)),
     q3=function(x)as.vector(quantile(x, probs=0.75)))

m1d()

nav <- l_ng_plots(m1d, color=iris$Species)

# with only one measure
nav <- l_ng_plots(measures1d(oliveAcids, sd))

# with two measures
nav <- l_ng_plots(measures1d(oliveAcids, sd=sd, mean=mean))
}

