% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ng_plots.R
\name{l_ng_plots.scagnostics}
\alias{l_ng_plots.scagnostics}
\title{Select 2d spaces in scatterplot matrix}
\usage{
\method{l_ng_plots}{scagnostics}(measures, data, separator = ":", ...)
}
\description{
Select 2d spaces in scatterplot matrix
}
\examples{

library(scagnostics)
scags <- scagnostics(oliveAcids)

l_ng_plots(scags, oliveAcids, color=olive$Area)
}

