% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ng_ranges.R
\name{l_ng_ranges.measures}
\alias{l_ng_ranges.measures}
\title{Navgraph with measures and sliders}
\usage{
\method{l_ng_ranges}{measures}(measures, ...)
}
\arguments{
\item{measures}{measure closure of class measures}

\item{...}{arguments get passed to scatterplot}
}
\description{
Navgraph with measures and sliders
}
\examples{


# 2d measures
s <- scagnostics2d(oliveAcids)
nav <- l_ng_ranges(s, color=olive$Area)

# 1d measures
scale01 <- function(x){(x-min(x))/diff(range(x))}
m1d <- measures1d(sapply(iris[,-5], scale01),
     mean=mean, median=median, sd=sd,
     q1=function(x)as.vector(quantile(x, probs=0.25)),
     q3=function(x)as.vector(quantile(x, probs=0.75)))

m1d()

nav <- l_ng_ranges(m1d, color=iris$Species)
}

