% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ng_ranges.R
\name{l_ng_ranges.scagnostics}
\alias{l_ng_ranges.scagnostics}
\title{Navigation Graphs based on Scagnostics Measures}
\usage{
\method{l_ng_ranges}{scagnostics}(measures, data, separator = ":", ...)
}
\description{
Navigation Graphs based on Scagnostics Measures
}
\examples{

library(scagnostics)
s <- scagnostics(oliveAcids)
ng <- l_ng_ranges(s, oliveAcids, color=olive$Area)
}

