% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer_map.R
\name{l_plot.map}
\alias{l_plot.map}
\title{Create a plot with a map layer}
\usage{
\method{l_plot}{map}(x, ...)
}
\arguments{
\item{x}{object of class map (defined in the maps library)}

\item{...}{arguments forwarded to \code{\link{l_layer.map}}}
}
\value{
Scatterplot widget plot handle
}
\description{
Creates a scatterplot widget and layers the map in front.
}
\examples{
if (requireNamespace("maps", quietly = TRUE)) {
   p <- l_plot(maps::map('world', fill=TRUE, plot=FALSE))
}
}
\seealso{
\code{\link{l_layer}}, \code{\link{l_layer.map}},
  \code{\link[maps]{map}}
}
