% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_plot.R
\name{l_plot}
\alias{l_plot}
\title{Create a 2d Scatterplot Widget}
\usage{
l_plot(x, y, ...)
}
\arguments{
\item{x}{coordinates}

\item{y}{coordinates}

\item{parent}{parent widget path of the scatterplot widget}
}
\value{
widget path name if was successful, otherwise NULL
}
\description{
Create a scatterplot widget. It the parent widget gets specified,
the plot widget needs to be placed using some geometry manager like
\code{\link{tkpack}} or \code{\link{tkplace}} in order to be
displayed.
}
\details{
This function is like \code{loon_plot} except it requires a parent
widget and need to be packed manually. This is useful when placing
more than one scatterplot in a window, or writing a custom Tk GUI
with a scatterplot.
}
\examples{

tt <- l_toplevel()
p1 <- loon_plot(parent=tt, x=c(1,2,3), y=c(3,2,1))
p2 <- loon_plot(c(4,3,1), (6,8,4), tt)

tkgrid(p1, row=0, column=0, sticky="nesw")
tkgrid(p2, row=0, column=1, sticky="nesw")
tkgrid.columnconfigure(tt, 0, weight=1)
tkgrid.columnconfigure(tt, 1, weight=1)
tkgrid.rowconfigure(tt, 0, weight=1)

tktitle(tt) <- "Loon plots with custom layout"
}

