% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{l_plot_inspector}
\alias{l_plot_inspector}
\title{Create a Scatterplot Inspector}
\usage{
l_plot_inspector(parent = NULL, ...)
}
\arguments{
\item{parent}{parent widget path}

\item{...}{state arguments}
}
\value{
widget handle
}
\description{
Inpectors provide graphical user interfaces to oversee and
  modify plot states
}
\examples{
if(interactive()){

i <- l_plot_inspector()
}
}
\seealso{
\code{\link{l_create_handle}}
}
