% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_predict.R
\name{l_predict}
\alias{l_predict}
\alias{l_predict.lm}
\alias{l_predict.nls}
\alias{l_predict.glm}
\alias{l_predict.loess}
\title{Model Prediction}
\usage{
l_predict(model, ...)

\method{l_predict}{lm}(
  model,
  newdata = NULL,
  interval = c("confidence", "none", "prediction"),
  level = 0.95,
  ...
)

\method{l_predict}{nls}(
  model,
  newdata = NULL,
  interval = c("confidence", "none", "prediction"),
  level = 0.95,
  ...
)

\method{l_predict}{glm}(
  model,
  newdata = NULL,
  interval = c("confidence", "none", "prediction"),
  level = 0.95,
  ...
)

\method{l_predict}{loess}(
  model,
  newdata = NULL,
  interval = c("confidence", "none", "prediction"),
  level = 0.95,
  ...
)
}
\arguments{
\item{model}{a model object for which prediction is desired}

\item{...}{arguments passed in \code{predict}}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict.
If omitted, the fitted linear predictors are used.}

\item{interval}{type of interval, could be "none", "confidence" or "prediction" (not for \code{glm})}

\item{level}{confidence level}
}
\value{
A data frame is returned with \code{x} (if \code{newdata} is given) and \code{y}.
If the \code{interval} is not \code{none},
two more columns, \code{lower} (lower interval) and \code{upper} (upper interval) will be returned.
}
\description{
It is entirely for the purpose of plotting fits and intervals on a scatterplot (or histogram).
It is a generic function to predict models for \code{loon} smooth layer (a wrap of the function \code{predict}).
However, the output is unified.
}
\examples{
y <- rnorm(10)
x <- rnorm(10)
model1 <- lm(y ~ x)
pre <- l_predict(model1, newdata = data.frame(x = sort(x)),
                 interval = "conf")

model2 <- loess(dist ~ speed, cars)
pre <- l_predict(model2,
                 newdata = data.frame(speed = sort(cars$speed)),
                 interval = "pred")

}
