% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_redraw.R
\name{l_redraw}
\alias{l_redraw}
\title{Force a Content Redraw of a Plot}
\usage{
l_redraw(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\description{
Force redraw the plot to make sure that all the visual elements 
  are placed correctly.
}
\details{
Note that this function is intended for debugging. If you find that 
  the display does not display the data according to its plot states then
  please contact loon's package maintainer.
}
\examples{
p <- l_plot(iris)
l_redraw(p)
}
