% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_resize.R
\name{l_resize}
\alias{l_resize}
\title{Resize Plot Widget}
\usage{
l_resize(widget, width, height)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{width}{width in pixels}

\item{height}{in pixels}
}
\description{
Resizes the toplevel widget to a specific size.
}
\examples{
p <- l_plot(iris)

l_resize(p, 300, 300)
l_size(p) <- c(500, 500)
}
\seealso{
\code{\link{l_size}}, \code{\link{l_size<-}}
}
