% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedLists.R
\name{l_Rlist2nestedTclList}
\alias{l_Rlist2nestedTclList}
\title{Convert an R list to a nested tcl list}
\usage{
l_Rlist2nestedTclList(x)
}
\description{
Convert an R list to a nested tcl list
}
\examples{

x <- list(c(1,3,4), c(4,3,2,1), c(4,3,2,5,6))
l_Rlist2nestedTclList(x)
}

