% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_scaleto.R
\name{l_scaleto_layer}
\alias{l_scaleto_layer}
\title{Change Plot Region to Display All Elements of a Particular Layer}
\usage{
l_scaleto_layer(target, layer)
}
\arguments{
\item{layer}{layer id}

\item{widget}{widget path as a string or as an object handle}
}
\description{
The function modifies the \code{zoomX}, \code{zoomY}, 
  \code{panX}, and \code{panY} so that all elements of a particular layer are
  displayed.
}
\seealso{
\code{\link{l_layer_ids}}
}

