% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_scaleto.R
\name{l_scaleto_selected}
\alias{l_scaleto_selected}
\title{Change Plot Region to Display All Selected Data}
\usage{
l_scaleto_selected(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\description{
The function modifies the \code{zoomX}, \code{zoomY},
  \code{panX}, and \code{panY} so that all selected data points are
  displayed.
}
