% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_scaleto.R
\name{l_scaleto_world}
\alias{l_scaleto_world}
\title{Change Plot Region to Display All Plot Data}
\usage{
l_scaleto_world(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\description{
The function modifies the \code{zoomX}, \code{zoomY}, 
  \code{panX}, and \code{panY} so that all elements in the plot are
  displayed.
}
