% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_serialaxes.R
\name{l_serialaxes}
\alias{l_serialaxes}
\title{Create an interactive serialaxes (parallel axes or radial axes) plot}
\usage{
l_serialaxes(data, ...)
}
\arguments{
\item{data}{a data frame with numerical data only}

\item{...}{named arguments to modify the serialaxes plot states or layouts}
}
\value{
widget handle
}
\description{
\code{l_serialaxes} is a generic function for creating interactive visualization environments for R objects.
}
\details{
For more information run: \code{l_help("learn_R_display_hist")}
}
\examples{
if(interactive()) {
   s <- l_serialaxes(data=oliveAcids, color=olive$Area, title="olive data")
   s['axesLayout'] <- 'parallel'
   states <- l_info_states(s)
   names(states)
}
}
\seealso{
\code{\link{l_serialaxes.default}}
}
