% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ColorList.R
\name{l_setColorList_ColorBrewer}
\alias{l_setColorList_ColorBrewer}
\title{ColorBrewer}
\usage{
l_setColorList_ColorBrewer(palette = c("Set1", "Set2", "Set3", "Pasetl1",
  "Pastel2", "Paired", "Dark2", "Accent"))
}
\description{
ColorBrewer
}
\examples{
l_setColorList_ColorBrewer("Set1")
p <- l_plot(iris)
}

