% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ColorList.R
\name{l_setColorList_ggplot2}
\alias{l_setColorList_ggplot2}
\title{Set loon's color mapping list to the colors from ggplot2}
\usage{
l_setColorList_ggplot2()
}
\description{
Loon's color list is used to map nominal values to colors. See
  the documentation for \code{\link{l_setColorList}}.
}
\seealso{
\code{\link{l_setColorList}}, \code{\link{l_setColorList_loon}},
  \code{\link{l_setColorList_ColorBrewer}}, \code{\link{l_setColorList_hcl}},
  \code{\link{l_setColorList_baseR}}, \code{\link{l_setColorList_ggplot2}}
}
