% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getOption.R
\name{l_setOption}
\alias{l_setOption}
\title{Set the value of a loon display option}
\usage{
l_setOption(option, value)
}
\arguments{
\item{option}{the name of the option being set}

\item{value}{the value to be assigned to the option.  If value == "default",
then the option is set to loon's default value for it.}
}
\value{
the new value
}
\description{
All of loon's displays access a set of common options.
This function assigns the value to the named option.
}
\examples{
l_setOption("select-color", "red")
l_setOption("select-color", "default")

}
\seealso{
\code{\link{l_getOption}}, \code{\link{l_getOptionNames}}, \code{\link{l_userOptions}},
  \code{\link{l_userOptionDefault}}
}
