% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_setTitleFont.R
\name{l_setTitleFont}
\alias{l_setTitleFont}
\title{Change all title fonts of loon's displays}
\usage{
l_setTitleFont(size = "16", weight = "bold", family = "Helvetica")
}
\description{
Change all title fonts of loon's displays
}
\examples{
loon:::l_setTitleFont(14)
}

