% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_size.R
\name{l_size<-}
\alias{l_size<-}
\title{Resize Plot Widget}
\usage{
l_size(widget) <- value
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{value}{numeric vector of length 2 with width and height in pixels}
}
\description{
Resizes the toplevel widget to a specific size. This setter
  function uses \code{\link{l_resize}}.
}
\examples{
if(interactive()){

p <- l_plot(iris)

l_resize(p, 300, 300)
l_size(p) <- c(500, 500)

}
}
\seealso{
\code{\link{l_resize}}, \code{\link{l_size}}
}
