% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_size.R
\name{l_size}
\alias{l_size}
\title{Query Size of a Plot Display}
\usage{
l_size(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
Vector width width and height in pixels
}
\description{
Get the width and height of a plot in pixels
}
\seealso{
\code{\link{l_resize}}, \code{\link{l_size<-}}
}
