% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_throwErrorIfNotLoonWidget.R
\name{l_throwErrorIfNotLoonWidget}
\alias{l_throwErrorIfNotLoonWidget}
\title{Throw an error if string is not associated with a loon widget}
\usage{
l_throwErrorIfNotLoonWidget(widget)
}
\arguments{
\item{widget}{widget path name as a string}
}
\value{
TRUE if the string is associated with a loon widget,
otherwise an error is thrown.
}
\description{
Helper function to ensure that a widget path is associated with a
loon widget.
}
