% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getOption.R
\name{l_userOptionDefault}
\alias{l_userOptionDefault}
\title{Get loon's system default value for the named display option.}
\usage{
l_userOptionDefault(option)
}
\arguments{
\item{option}{the name of the user changeable loon display option 
whose default value is to be determined.}
}
\value{
the default value for the named option
}
\description{
All of loon's displays access a set of common options.  
This function accesses and returns the default value for the named option.
}
\examples{
l_userOptionDefault("background")

}
\seealso{
\code{\link{l_getOptionNames}}, \code{\link{l_getOption}},
  \code{\link{l_userOptionDefault}}, \code{\link{l_userOptions}}
}
