% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_getOption.R
\name{l_userOptions}
\alias{l_userOptions}
\title{Get the names of all loon display options that can be set by the user.}
\usage{
l_userOptions()
}
\value{
a vector of all user settable option names.
}
\description{
All of loon's displays access a set of common options.
This function accesses and returns the names of the subset of loon options which
can be changed by the user.
}
\examples{
l_userOptions()

}
\seealso{
\code{\link{l_getOptionNames}}, \code{\link{l_getOption}},
  \code{\link{l_userOptionDefault}}, \code{\link{l_setOption}}
}
