% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_widget.R
\name{l_widget}
\alias{l_widget}
\title{Dummy function to be used in the Roxygen documentation}
\usage{
l_widget(widget)
}
\arguments{
\item{widget}{widget path name as a string}
}
\value{
widget path name as a string
}
\description{
Dummy function to be used in the Roxygen documentation
}
