% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_zoom.R
\name{l_zoom}
\alias{l_zoom}
\title{Zoom from and towards the center}
\usage{
l_zoom(widget, factor = 1.1)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{factor}{a zoom factor}
}
\description{
This function changes the plot states \code{panX}, \code{panY}, 
  \code{zoomX}, and \code{zoomY} to zoom towards or away from the center of
  the current view.
}
