% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphutils.R
\name{linegraph.loongraph}
\alias{linegraph.loongraph}
\title{Create a linegraph of a graph}
\usage{
\method{linegraph}{loongraph}(x, separator = ":", ...)
}
\arguments{
\item{x}{loongraph object}

\item{separator}{one character - node names in x get concatenated with this 
character}

\item{...}{additional arguments are not used for this methiod}
}
\value{
graph object of class loongraph
}
\description{
Create a lingraph of a loongraph
}
\details{
linegraph.loongraph needs the code part for directed graphs (i.e.
  isDirected=TRUE)
}
\examples{
g <- loongraph(letters[1:4], letters[1:3], letters[2:4], FALSE)

linegraph(g)
}
