% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_layer_scatterplot.R
\name{loonGlyphGrob}
\alias{loonGlyphGrob}
\title{Create a grob glyph from a loon widget}
\usage{
loonGlyphGrob(widget, x, glyph_info)
}
\arguments{
\item{widget}{the loon widget.}

\item{x}{argument used to dispatch the method -- an empty structure of class equal to that
returned by \code{l_glyph_getType}.}

\item{glyph_info}{a named list of pertinent components of the glyph including
its x and y locations in the plot as well as other information relevant to the
particular glyph.}
}
\value{
A grob for that glyph.
}
\description{
A generic function used by \code{loonGrob} specialized for particular loon widgets.
Used to construct the various point symbol types of the plot.
Different \code{S3} methods are implemented for various loon point glyphs.
}
\seealso{
\code{\link{loonGrob}}
}
