% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob.R
\name{loonGrob}
\alias{loonGrob}
\title{Create a grid grob from a loon widget handle}
\usage{
loonGrob(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
a grid grob
}
\description{
Grid grobs are useful to create publication quality graphics.
}
\examples{

library(grid)
l_setColorList_baseR()

widget <- with(iris, l_plot(Sepal.Length, Sepal.Width))

g <- loonGrob(widget)
grid.ls(g, viewports=TRUE, fullNames=TRUE)
grid.newpage()
grid.draw(g)

widget['title'] <- "Iris Data"

grid.edit("Scatterplot", gp=gpar(alpha = 0.6))

widget['showScales'] <- TRUE
g <- loonGrob(widget)
grid.newpage()
grid.draw(g)


}
