% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures1d}
\alias{measures1d}
\title{One Dimensional Measures}
\usage{
measures1d(data, ...)
}
\description{
One Dimensional Measures
}
\examples{
m1 <- measures1d(oliveAcids, mean=mean, median=median,
     sd=sd, q1=function(x)as.vector(quantile(x, probs=0.25)),
     q3=function(x)as.vector(quantile(x, probs=0.75)))
     
m1
m1()
m1(olive$palmitoleic>100)
m1('data')
m1('measures')
}

