% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures2d}
\alias{measures2d}
\title{Two Dimensional Measures}
\usage{
measures2d(data, ...)
}
\description{
Two Dimensional Measures
}
\examples{
m <- measures2d(oliveAcids, separator='*', cov=cov, cor=cor)
m
m()
m(keep=olive$palmitic>1360)
m('data')
m('grid')
m('measures')
}

