% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_state_names.R
\name{names.loon}
\alias{names.loon}
\alias{{l_state_names}}
\title{Get State Names of Loon Object}
\usage{
\method{names}{loon}(x)
}
\arguments{
\item{x}{loon object}
}
\value{
state names
}
\description{
States of loon objects can be accessed \code{`[`} and \code{\link{l_cget}}
and modified with \code{\link{l_configure}}.
}
\seealso{
Other loon interactive states: 
\code{\link{l_info_states}()},
\code{\link{l_plot.default}()},
\code{\link{l_plot}()},
\code{\link{l_state_names}()}
}
\concept{loon interactive states}
