% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_nDimStateNames.R
\name{nDimStateNames}
\alias{nDimStateNames}
\alias{l_nDimStateNames}
\title{N dimensional state names access}
\usage{
l_nDimStateNames(loon_plot)
}
\arguments{
\item{loon_plot}{A \code{loon} widget or a specified character}
}
\description{
Get, set, delete n dimensional state names
}
\examples{
p <- l_plot()
l_nDimStateNames(p)
l_nDimStateNames("l_plot")
\dontrun{
  l_set_nDimStateNames <- function(name, key) {
    if(missing(name)) stop("Name is missing")
    if(missing(key)) stop("Key is missing")
    envir <- getNamespace("loon")
    name <- as.character(name)
    n_dim_states_list <- get("n_dim_states_list", envir = envir)
    n_dim_states_list <- c(n_dim_states_list,
                           setNames(list(key), name))
    assignInNamespace(x = "n_dim_states_list",
                      value = n_dim_states_list,
                      ns = envir)
    invisible()
 }

 l_delete_nDimStateNames <- function(name) {

    if(missing(name)) stop("Name is missing")

    name <- as.character(name)
    envir <- getNamespace("loon")

    n_dim_states_list <- get("n_dim_states_list", envir = envir)
    if(!name \%in\% c("l_plot.default", "l_hist", "l_graph", "l_serialaxes",
                    "l_plot3D", "l_layer_lines", "l_layer_polygons", "l_layer_texts",
                    "l_layer_rectangles", "l_layer_points")) {
        n_dim_states_list[name] <- NULL
        assignInNamespace(x = "n_dim_states_list",
                          value = n_dim_states_list,
                          ns = envir)
    } else {
        warning(
            "N dimension state names of widget ",
            name,
            " are not allowed to be modified" ,
            call. = FALSE)
    }
    invisible()
}

## Suppose l_foo is the function you created
## Its n dimensional state is c("x", "y", "linkingKey")
l_nDimStateNames("l_foo")
l_set_nDimStateNames("l_foo", c("x", "y", "linkingKey"))
l_nDimStateNames("l_foo")
l_delete_nDimStateNames("l_foo")
}
}
