% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oliveAcids}
\alias{oliveAcids}
\title{Just the Fatty Acid Composition of Italian Olive Oils}
\format{A data frame containing 572 cases and 8 variates.}
\usage{
oliveAcids
}
\description{
This is the \code{\link{olive}} data set minus the \code{Region}
  and \code{Area} variables.
}
\seealso{
\code{\link{olive}}
}
\keyword{datasets}
