% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oliveLocations}
\alias{oliveLocations}
\title{Locations of olive growing regions in Italy.

A dataset containing the latitude and longitude of the nine growing
regions from the olive data set.}
\format{A data frame with 9 rows and 9 variables:}
\source{
\url{https://www.latlong.net}
}
\usage{
oliveLocations
}
\description{
Locations of olive growing regions in Italy.

A dataset containing the latitude and longitude of the nine growing
regions from the olive data set.
}
\seealso{
\code{\link{olive}}
}
\keyword{datasets}
