% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphutils.R
\name{plot.loongraph}
\alias{plot.loongraph}
\title{Plot a loon graph object with base R graphics}
\usage{
\method{plot}{loongraph}(x, ...)
}
\arguments{
\item{x}{object of class loongraph}

\item{...}{arguments forwarded to method}
}
\description{
This function converts the loongraph object to one of class 
  graph and the plots it with its respective plot method.
}
\examples{
if (requireNamespace("Rgraphviz", quietly = TRUE)) {
  g <- loongraph(letters[1:4], letters[1:3], letters[2:4], FALSE)
  plot(g)
}
}
