% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_generics.R
\name{print.l_layer}
\alias{print.l_layer}
\title{Print a summary of a loon layer object}
\usage{
\method{print}{l_layer}(x, ...)
}
\arguments{
\item{x}{an \code{l_layer} object}

\item{...}{additional arguments are not used for this methiod}
}
\description{
Prints the layer label and layer type
}
\seealso{
\code{\link{l_layer}}
}
