% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{print.measures1d}
\alias{print.measures1d}
\title{Print function names from measure1d object}
\usage{
\method{print}{measures1d}(x, ...)
}
\arguments{
\item{x}{measures1d object}

\item{...}{arguments passed on to print.default}
}
\description{
Prints the function names of a measure1d object using
  \code{print.default}.
}
