% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{print.measures2d}
\alias{print.measures2d}
\title{Print function names from measure2d object}
\usage{
\method{print}{measures2d}(x, ...)
}
\arguments{
\item{x}{measures2d object}

\item{...}{arguments passed on to print.default}
}
\description{
Prints the function names of a measure2d object using
  \code{print.default}.
}
