% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scagnostics2d.R
\name{scagnostics2d}
\alias{scagnostics2d}
\title{Created 2d scagnostics measure closure}
\usage{
scagnostics2d(data, scagnostics = c("Clumpy", "Monotonic", "Convex",
  "Stringy", "Skinny", "Outlying", "Sparse", "Striated", "Skewed"),
  separator = ":")
}
\description{
We have added Not... to all the scagnostics measures.
}
\examples{
m <- scagnostics2d(oliveAcids, separator='**')
m
m()
m(olive$palmitoleic > 80)
m('data')
m('grid')
m('measures')
}

