% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer_sp.R
\name{spAsList}
\alias{spAsList}
\title{Create a list of polygons or lines}
\usage{
spAsList(x)
}
\description{
Create a list of polygons or lines
}
\examples{
 
library(sp)
library(rworldmap)
world <- getMap(resolution = "coarse")
class(world)
isS4(world)
xy <- spAsList(world)

names(xy)
# because tree has same depth for every leaf unlist is ok
uxy <- spunlist(xy)
unlist(xy, recursive=FALSE)

names(uxy)

# but here, unlist would be wrong
a <- list(list(x=1:2, y=1:2), list(list(x=1:3, y=1:3), list(x=1:4, y=1:4)))
spunlist(a)
unlist(a, recursive=FALSE)
}

