% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob_l_layer_scatterplot.R
\name{tcl_img_2_r_raster}
\alias{tcl_img_2_r_raster}
\title{A tk Image Object to a Raster Object}
\usage{
tcl_img_2_r_raster(img)
}
\arguments{
\item{img}{a tk image object}
}
\description{
Turn a tk image object to an R \code{raster} object
}
\examples{
if(requireNamespace("grid")) {
puglia <- list.files(file.path(find.package(package = 'loon'), "images"),
                     full.names = TRUE)[1L]
# `img` is a tk image object
img <- setNames(l_image_import_files(puglia),
                tools::file_path_sans_ext(basename(puglia)))
raster <- tcl_img_2_r_raster(img)
grid::grid.newpage()
grid::grid.raster(raster)
}
}
