% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UsAndThem}
\alias{UsAndThem}
\title{Data to re-create Hans Rosling's fameous "Us and Them" animation}
\format{A data frame with 9855 rows and 8 variables}
\source{
\url{http://www.gapminder.org/}
}
\usage{
UsAndThem
}
\description{
This data was sourced from \url{https://www.gapminder.org/} and
  contains Population, Life Expectancy, Fertility, Income, and
  Geographic.Region information between 1962 and 2013 for 198 countries.
}
\keyword{datasets}
