% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ColorList.R
\name{hex12tohex6}
\alias{hex12tohex6}
\title{Convert 12 hexadecimal digit color representations to 6 hexidecimal digit
color representations}
\usage{
hex12tohex6(x)
}
\arguments{
\item{x}{a vector with 12 digit hexcolors}
}
\description{
Tk colors must be in 6 hexadecimal format with two hexadecimal
digits for each of the red, green, and blue components.  Twelve hexadecimal digit
colors have 4 hexadecimal digits for each.  This function converts the 12 digit format to the 6
provided the color is preserved.
}
\details{
Function throws a warning if the conversion loses information. The
  \code{\link{l_hexcolor}} function converts any Tcl color specification to a
  12 digit hexadecimal color representation.
}
\examples{
x <- l_hexcolor(c("red", "green", "blue", "orange"))
x
hex12tohex6(x)

}
