% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_bind.R
\name{l_bind_state_delete}
\alias{l_bind_state_delete}
\title{Delete a state binding}
\usage{
l_bind_state_delete(target, id)
}
\arguments{
\item{target}{either an object of class loon or a vector that specifies the
widget, layer, glyph, navigator or context completely. The widget is
specified by the widget path name (e.g. \code{'.l0.plot'}), the remaining
objects by their ids.}

\item{id}{state binding id}
}
\description{
Remove a state binding
}
\details{
Bindings, callbacks, and binding substitutions are described in detail in
loon's documentation webpage, i.e. run \code{l_help("learn_R_bind")}
}
\seealso{
\code{\link{l_bind_state}}, \code{\link{l_bind_state_ids}},
  \code{\link{l_bind_state_get}}, \code{\link{l_bind_state_reorder}}
}
