% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_compound.R
\name{l_getLocations}
\alias{l_getLocations}
\title{For the target compound loon plot, determines location (only and excluding the grobs)
arguments to pass to `gridExtra::arrangeGrob()`}
\usage{
l_getLocations(target)
}
\arguments{
\item{target}{the (compound) loon  plot whose locations are needed lay it out.}
}
\value{
a list of an appropriate subset of the named location arguments
`c("ncol", "nrow", "layout_matrix", "heights", "widths")` and their values
 to be passed to `gridExtra::arrangeGrob()`.
}
\description{
For the target compound loon plot, determines location (only and excluding the grobs)
arguments to pass to `gridExtra::arrangeGrob()`
}
