% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_compound.R
\name{l_getPlots}
\alias{l_getPlots}
\title{For the target compound loon plot, determines all the loon plots
in that compound plot.}
\usage{
l_getPlots(target)
}
\arguments{
\item{target}{the (compound) loon  plot to be laid out.}
}
\value{
a list of the named arguments and their values to be passed to
`gridExtra::arrangeGrob()`.
}
\description{
For the target compound loon plot, determines all the loon plots
in that compound plot.
}
