% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_glyphs.R
\name{l_glyph_add.default}
\alias{l_glyph_add.default}
\title{Default method for adding non-primitive glyphs}
\usage{
\method{l_glyph_add}{default}(widget, type, label = "", ...)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{type}{loon-native non-primitive glyph type, one of \code{'text'},
\code{'serialaxes'}, \code{'image'}, \code{'[polygon'}, or
\code{'pointrange'}}

\item{label}{label of a glyph (currently shown only in the glyph inspector)}

\item{...}{state arguments}
}
\description{
Generic function to write new glyph types using loon's primitive
  glyphs
}
\seealso{
Other glyph functions: \code{\link{l_glyph_add_image}},
  \code{\link{l_glyph_add_pointrange}},
  \code{\link{l_glyph_add_polygon}},
  \code{\link{l_glyph_add_serialaxes}},
  \code{\link{l_glyph_add_text}},
  \code{\link{l_glyph_add}}, \code{\link{l_glyph_delete}},
  \code{\link{l_glyph_getLabel}},
  \code{\link{l_glyph_getType}}, \code{\link{l_glyph_ids}},
  \code{\link{l_glyph_relabel}},
  \code{\link{l_primitiveGlyphs}}
}
\concept{glyph functions}
