% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_glyphs.R
\name{l_glyph_getLabel}
\alias{l_glyph_getLabel}
\title{Get Glyph Label}
\usage{
l_glyph_getLabel(widget, id)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{id}{glyph id}
}
\description{
Returns the label of a glyph
}
\seealso{
\code{\link{l_glyph_add}}, \code{\link{l_glyph_ids}},
  \code{\link{l_glyph_relabel}}

Other glyph functions: \code{\link{l_glyph_add.default}},
  \code{\link{l_glyph_add_image}},
  \code{\link{l_glyph_add_pointrange}},
  \code{\link{l_glyph_add_polygon}},
  \code{\link{l_glyph_add_serialaxes}},
  \code{\link{l_glyph_add_text}},
  \code{\link{l_glyph_add}}, \code{\link{l_glyph_delete}},
  \code{\link{l_glyph_getType}}, \code{\link{l_glyph_ids}},
  \code{\link{l_glyph_relabel}},
  \code{\link{l_primitiveGlyphs}}
}
\concept{glyph functions}
