% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_glyphs.R
\name{l_glyph_getType}
\alias{l_glyph_getType}
\title{Get Glyph Type}
\usage{
l_glyph_getType(widget, id)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{id}{glyph id}
}
\description{
Query the type of a glyph
}
\seealso{
\code{\link{l_glyph_add}}
}
