% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_glyphs.R
\name{l_glyph_relabel}
\alias{l_glyph_relabel}
\title{Relabel Glyph}
\usage{
l_glyph_relabel(widget, id, label)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}

\item{id}{glyph id}

\item{label}{new label}
}
\description{
Change the label of a glyph. Note that the label is only
  displayed in the glyph inspector.
}
\examples{
if(interactive()){

p <- l_plot(iris, color = iris$Species)
g <- l_glyph_add_text(p, iris$Species, "test_label")
p['glyph'] <- g
l_glyph_relabel(p, g, "Species")
}
}
\seealso{
Other glyph functions: 
\code{\link{l_glyph_add.default}()},
\code{\link{l_glyph_add_image}()},
\code{\link{l_glyph_add_pointrange}()},
\code{\link{l_glyph_add_polygon}()},
\code{\link{l_glyph_add_serialaxes}()},
\code{\link{l_glyph_add_text}()},
\code{\link{l_glyph_add}()},
\code{\link{l_glyph_delete}()},
\code{\link{l_glyph_getLabel}()},
\code{\link{l_glyph_getType}()},
\code{\link{l_glyph_ids}()},
\code{\link{l_primitiveGlyphs}()}
}
\concept{glyph functions}
