% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_graphswitch.R
\name{l_graphswitch_get}
\alias{l_graphswitch_get}
\title{Return a Graph as a loongraph Object}
\usage{
l_graphswitch_get(widget, id)
}
\arguments{
\item{widget}{graphswitch widget handle (or widget path)}

\item{id}{of the graph}
}
\description{
Graphs can be extracted from the graphswitch widget as loongraph
  objects.
}
\seealso{
\code{\link{l_graphswitch}}, \code{\link{loongraph}}
}
