% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_help.R
\name{l_help}
\alias{l_help}
\title{Open a browser with loon's combined (TCL and R) documentation website}
\usage{
l_help(page = "index", ...)
}
\arguments{
\item{page}{relative path to a page, the .html part may be omitted}

\item{...}{arguments forwarded to browseURL, e.g. to specify a browser}
}
\description{
\code{l_help} opens a browser with the relevant page on the
  official combined loon documentation website at
  \url{https://great-northern-diver.github.io/loon/l_help/}.
}
\examples{

\dontrun{
l_help()
l_help("learn_R_intro")
l_help("learn_R_display_hist")
l_help("learn_R_bind")
# jump to a section
l_help("learn_R_bind.html#list-reorder-delete-bindings")
}
}
\seealso{
\code{\link{help}}, \code{\link{l_web}} for \code{R} manual or
web \code{R} manual
}
