% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_layer_smooth.R
\name{l_layer_smooth}
\alias{l_layer_smooth}
\title{layer a 2d smooth}
\usage{
l_layer_smooth(widget, method = c("lm", "glm", "gam", "loess", "rlm"),
  formula = y ~ x, loon.args = list(fit, confidence), ...)
}
\examples{

p <- l_plot(iris)

l_layer_smooth(p)

l_layer_smooth(p, method="lm", formula=y~x+x^2+2)


fit <- lm(dist ~ speed, data=cars)

head(predict(fit, interval=c("confidence", "prediction")))


}

