% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{l_loon_inspector}
\alias{l_loon_inspector}
\title{Create a loon inspector}
\usage{
l_loon_inspector(parent = NULL, ...)
}
\arguments{
\item{parent}{parent widget path}

\item{...}{state arguments, see \code{\link{l_info_states}}.}
}
\value{
widget handle
}
\description{
The loon inspector is a singleton widget that provids an
  overview to view and modify the active plot.
}
\details{
For more information run: \code{l_help("learn_R_display_inspectors")}
}
\examples{
i <- l_loon_inspector()
}
