% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_pairs.R
\name{l_pairs}
\alias{l_pairs}
\title{Scatterplot Matrix in Loon}
\usage{
l_pairs(data, parent = NULL, ...)
}
\arguments{
\item{data}{a data.frame with numerical data to create the scatterplot matrix}

\item{parent}{parent widget path}

\item{...}{named arguments to modify the scatterplot states}
}
\value{
a list with scatterplot handles
}
\description{
Function creates a scatterplot matrix using loon's scatterplot 
  widgets
}
\examples{
p <- l_pairs(iris[,-5], color=iris$Species)
}
\seealso{
\code{\link{l_plot}}
}
