% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedLists.R
\name{l_Rlist2nestedTclList}
\alias{l_Rlist2nestedTclList}
\title{Convert an R list to a nested Tcl list}
\usage{
l_Rlist2nestedTclList(x)
}
\arguments{
\item{x}{a list of vectors}
}
\value{
a string that represents the tcl nested list
}
\description{
This is a helper function to create a nested Tcl list from an R
  list (i.e. a list of vectors).
}
\examples{
x <- list(c(1,3,4), c(4,3,2,1), c(4,3,2,5,6))
l_Rlist2nestedTclList(x)
}
\seealso{
\code{\link{l_nestedTclList2Rlist}}
}
