% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_scaleto.R
\name{l_scaleto_layer}
\alias{l_scaleto_layer}
\title{Change Plot Region to Display All Elements of a Particular Layer}
\usage{
l_scaleto_layer(target, layer)
}
\arguments{
\item{target}{either an object of class loon or a vector that specifies the
widget, layer, glyph, navigator or context completely. The widget is
specified by the widget path name (e.g. \code{'.l0.plot'}), the remaining
objects by their ids.}

\item{layer}{layer id}
}
\description{
The function modifies the \code{zoomX}, \code{zoomY},
  \code{panX}, and \code{panY} so that all elements of a particular layer are
  displayed.
}
\seealso{
\code{\link{l_layer_ids}}
}
