% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_setTitleFont.R
\name{l_setTitleFont}
\alias{l_setTitleFont}
\title{Set the title font of all loon displays}
\usage{
l_setTitleFont(size = "16", weight = "bold", family = "Helvetica")
}
\arguments{
\item{size}{the font size.}

\item{weight}{the font size.}

\item{family}{the font family.}
}
\value{
the value of the named option.
}
\description{
All of loon's displays access a set of common options.
This function sets the font for the title bar of the displays.
}
\seealso{
\code{\link{l_getOptionNames}}, \code{\link{l_userOptions}},
  \code{\link{l_userOptionDefault}}, \code{\link{l_setOption}}
}
