% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_state_names.R
\name{l_state_names}
\alias{l_state_names}
\alias{{names.loon}}
\title{Get State Names of Loon Object}
\usage{
l_state_names(target)
}
\arguments{
\item{target}{either an object of class loon or a vector that specifies the
widget, layer, glyph, navigator or context completely. The widget is
specified by the widget path name (e.g. \code{'.l0.plot'}), the remaining
objects by their ids.}
}
\value{
state names
}
\description{
States of loon objects can be accessed \code{`[`} and \code{\link{l_cget}}
and modified with \code{\link{l_configure}}.
}
\details{
In order to access values of a states use \code{\link{l_info_states}}.
}
\seealso{
\code{\link{l_info_states}}, \code{\link{l_cget}}, \code{\link{l_configure}}

Other loon interactive states: \code{\link{l_info_states}},
  \code{\link{l_plot.default}}, \code{\link{l_plot}},
  \code{\link{names.loon}}
}
\concept{loon interactive states}
