% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loonGrob.R
\name{loonGrob}
\alias{loonGrob}
\title{Create a grid grob from a loon widget handle}
\usage{
loonGrob(widget)
}
\arguments{
\item{widget}{widget path as a string or as an object handle}
}
\value{
a grid grob
}
\description{
Grid grobs are useful to create publication quality graphics.
}
\examples{

library(grid)

widget <- with(iris, l_plot(Sepal.Length, Sepal.Width))

g <- loonGrob(widget)
grid.ls(g, viewports=TRUE, fullNames=TRUE)
grid.newpage()
grid.draw(g)

widget['title'] <- "Iris Data"

grid.edit("Scatterplot", gp=gpar(alpha = 0.6))

widget['showScales'] <- TRUE
g <- loonGrob(widget)
grid.newpage()
grid.draw(g)

grid.ls(g, viewports = TRUE, fullNames = TRUE)

## with primitive layers from l_layers demo

p <- with(olive,
          l_plot(x=linoleic, y=oleic,
                 color=Region, title="Olive Data"))
                 
## Layer a Group
l.g <- l_layer_group(p, label="Drawings", parent="root", index="end")

## Layer Points
l.pts <- l_layer_points(p,
                        x=seq(from=200,to=1600, length.out=20),
                        y=seq(from=6000,to=8000, length.out=20),
                        color="steelblue", size=20:39)

## Polygon
i <- with(olive, chull(linoleic, oleic))
x.hull <- olive$linoleic[i]
y.hull <- olive$oleic[i]
l_layer_polygon(p, x.hull, y.hull, color="thistle",
                linecolor="black", linewidth=4, parent=l.g)

## Rectangle
l_layer_rectangle(p, x=c(1100, 1300), y=c(7600, 8300), linewidth=2)

## Oval
l_layer_oval(p, x=c(1500, 1750), y=c(7900, 8100),
             color="", linecolor="orange", linewidth=4)

## Line
x <- with(olive, linoleic[Region=="North"])
y <- with(olive, oleic[Region=="North"])

fit <- lm(y~x)
xr <- seq(from=min(x), to=max(x), length.out=20)
yp <- predict(fit, data.frame(x=xr), interval="prediction")

l.pi <- l_layer_polygon(p, x=c(xr, rev(xr)),
                        y=c(yp[,2],rev(yp[,3])),
                        color="lightgreen",
                        linecolor= "darkgreen", linewidth=2,
                        label="predition interval west liguria")

l.fit <- l_layer_line(p, x=xr, y=yp[,1],
                      color="darkgreen", linewidth=8,
                      label="fit west liguria")


## Text (size does not work and color is gray)
bbox <- l_layer_bbox(p, "root")
l_layer_texts(p, x=seq(from=bbox[1], to=bbox[3], length.out=length(LETTERS)),
              y=rev(seq(from=bbox[2], to=bbox[4], length.out=length(LETTERS))),
              text=LETTERS, size=seq_along(LETTERS),
              angle=seq_along(LETTERS)/length(LETTERS)*360)

# text
l_layer_text(p, x = 750, y = 6500, text = "Hello World")

# polygons
l_layer_polygons(p,
  x = list(
    c(500, 500, 600, 750, 650, 550),
    c(500, 500, 750, 750, 625),
    c(1250, 1250, 1400, 1400),
    c(1250, 1250, 1400)
  ),
  y = list(
    c(6500, 6800, 7200, 6800, 6500, 7000),
    c(8200, 8500, 8500, 8200, 8400),
    c(8200, 8500, 8500, 8300),
    c(6500, 7200, 6850)
  ),
  color = c('blue', 'orange', 'yellow', 'red')
)

# rectangles
l_layer_rectangles(p,
  x = lapply(1:4, function(x) 500 + x*40 + c(0, 20)),
  y = lapply(1:4, function(y) 6500 + y*100 + c(0, 60))
)

# lines
l_layer_lines(p,
  x = replicate(3, c(1250, 1300, 1500), simplify = FALSE),
  y = list(
    c(7500, 8200, 8000),
    c(7200, 8000, 7200),
    c(6800, 6700, 7000)
  ),
  color = "magenta"
)

l_scaleto_world(p)

p['showScales'] <- TRUE

g <- loonGrob(p)
library(grid)
grid.newpage()
grid.draw(g)

grid.ls(g, viewports=TRUE, fullNames=TRUE)


}
