% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{olive}
\alias{olive}
\title{Fatty Acid Composition of Italian Olive Oils}
\format{A data frame containing 572 cases and 10 variates.}
\usage{
olive
}
\description{
This data set records the percentage composition of 8 fatty 
  acids (palmitic, palmitoleic, stearic, oleic, linoleic, linolenic, 
  arachidic, eicosenoic) found in the lipid fraction of 572 Italian olive 
  oils. The oils are samples taken from three Italian regions varying number 
  of areas within each region.  The regions and their areas are recorded as 
  shown in the following table:
  
  \tabular{ll}{ \bold{Region} \tab \bold{Area}\cr North \tab North-Apulia,
  South-Apulia, Calabria, Sicily \cr South \tab East-Liguria, West-Liguria,
  Umbria \cr Sardinia \tab Coastal-Sardinia, Inland-Sardinia }
}
\references{
Forina, M., Armanino, C., Lanteri, S., and Tiscornia, E. (1983)
  "Classification of Olive Oils from their Fatty Acid Composition", in Food
  Research and Data Analysis (Martens, H., Russwurm, H., eds.), p. 189,
  Applied Science Publ., Barking.
}
\keyword{datasets}
