% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oliveLocations}
\alias{oliveLocations}
\title{Geographic location of each Italian olive growing area
named in the \code{\link{olive}} data.}
\format{
A data frame containing 9 cases and 3 variates.
\describe{
 \item{Area}{name of the Italian growing area of the olive oil.}
 \item{lat, long}{latitude and longitude in degrees of the approximate
 centre of the named growing area}
}
}
\source{
\url{https://www.latlong.net}
}
\usage{
oliveLocations
}
\description{
A longitude and latitude for each \code{Area}
 named in the \code{\link{olive}} data set.
}
\seealso{
\code{\link{olive}}
}
\keyword{datasets}
